Covered all test cases for Dec Int in the test file, still needs to be formatted and cleaned up

+++ include/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: ahunt <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/20 09:22:06 by ahunt             #+#    #+#             */
-/*   Updated: 2016/10/25 03:37:38 by ahunt            ###   ########.fr       */
+/*   Updated: 2016/10/25 04:53:36 by ahunt            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,8 +24,8 @@ typedef struct	s_fmt
 	int			pos_val;
 	int			prepend_sp;
 	int			prepend_zeros;
-	size_t		width;
-	size_t		min_width;
+	int			width;
+	int			min_width;
 	int			length;
 }				t_fmt;
 
@@ -37,7 +37,7 @@ int				parse_conv_spec(va_list *ap, t_fmt **args, char **fmt);
 void			parse_flags(t_fmt **args, char **fmt);
 void			parse_num(t_fmt **args, char **fmt, int is_width);
 void			parse_length(t_fmt **args, char **fmt);
-size_t			ft_printf_int(va_list *ap, t_fmt **args);
+int				ft_printf_int(va_list *ap, t_fmt **args);
 intmax_t		ft_get_int_length(va_list *ap, t_fmt **args);
 int				get_number_length(int nbr);
